Move var-args and struct-by-value functions to libtrufflerubytrampoline
* This will allow rb-sys and magnus to reference such functions,
  see https://github.com/truffleruby/truffleruby/issues/3396#issuecomment-2580778963
