+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Compatibility:
 * `Rational#**` no longer returns a Float, it will raise an ArgumentError if the result is too big (#4105, @herwinw).
 * The dummy `Encoding::UTF_16` and `Encoding::UTF_32` encodings are now regular single-byte dummy encodings (#4083, @eregon).
 * Implement `rb_interned_str()` and `rb_interned_str_cstr()` (#4018, @herwinw).
+* Define C API var-args and struct-by-value functions as non-inline to support taking their address (#3396, @eregon).
 
 Performance:
