+++ lib/cext/include/ruby/internal/error.h
@@ -116,11 +116,6 @@ RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 2, 3)
  */
 void rb_raise(VALUE exc, const char *fmt, ...);
 
-#ifdef TRUFFLERUBY
-RBIMPL_ATTR_NORETURN()
-void rb_tr_fatal_va_list(const char *fmt, va_list args);
-#endif
-
 RBIMPL_ATTR_NORETURN()
 RBIMPL_ATTR_NONNULL((1))
 RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 1, 2)
@@ -133,21 +128,7 @@ RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 1, 2)
  * @exception  rb_eFatal  An exception that you cannot rescue.
  * @note       It never returns.
  */
-#ifdef TRUFFLERUBY
-static inline void rb_fatal(const char *fmt, ...) {
-    va_list args;
-    va_start(args, fmt);
-    rb_tr_fatal_va_list(fmt, args);
-    va_end(args);
-}
-#else
 void rb_fatal(const char *fmt, ...);
-#endif
-
-#ifdef TRUFFLERUBY
-RBIMPL_ATTR_NORETURN()
-void rb_tr_bug_va_list(const char *fmt, va_list args);
-#endif
 
 RBIMPL_ATTR_COLD()
 RBIMPL_ATTR_NORETURN()
@@ -169,16 +150,7 @@ RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 1, 2)
  * @param[in]  fmt  Format specifier string compatible with rb_sprintf().
  * @note       It never returns.
  */
-#ifdef TRUFFLERUBY
-static inline void rb_bug(const char *fmt, ...) {
-    va_list args;
-    va_start(args, fmt);
-    rb_tr_bug_va_list(fmt, args);
-    va_end(args);
-}
-#else
 void rb_bug(const char *fmt, ...);
-#endif
 
 RBIMPL_ATTR_NORETURN()
 RBIMPL_ATTR_NONNULL(())
@@ -513,10 +485,6 @@ VALUE *rb_ruby_debug_ptr(void);
  */
 #define ruby_debug   (*rb_ruby_debug_ptr())
 
-#ifdef TRUFFLERUBY
-void rb_tr_warn_va_list(const char *fmt, va_list args);
-#endif
-
 /* reports if $VERBOSE is true */
 RBIMPL_ATTR_NONNULL((1))
 RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 1, 2)
@@ -540,18 +508,7 @@ RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 1, 2)
  *
  * Above description is in fact inaccurate.  This API interfaces with Ractors.
  */
-#ifdef TRUFFLERUBY
-static inline void rb_warning(const char *fmt, ...) {
-    if (RTEST(ruby_verbose)) {
-        va_list args;
-        va_start(args, fmt);
-        rb_tr_warn_va_list(fmt, args);
-        va_end(args);
-    }
-}
-#else
 void rb_warning(const char *fmt, ...);
-#endif
 
 RBIMPL_ATTR_NONNULL((2))
 RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 2, 3)
@@ -598,23 +555,7 @@ RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 1, 2)
  *             nothing if $VERBOSE is nil.
  * @param[in]  fmt  Format specifier string compatible with rb_sprintf().
  */
-#ifdef TRUFFLERUBY
-static inline void rb_warn(const char *fmt, ...) {
-    if (!NIL_P(ruby_verbose)) {
-        va_list args;
-        va_start(args, fmt);
-        rb_tr_warn_va_list(fmt, args);
-        va_end(args);
-    }
-}
-#else
 void rb_warn(const char *fmt, ...);
-#endif
-
-#ifdef TRUFFLERUBY
-bool rb_warning_category_enabled_p(rb_warning_category_t category);
-void rb_tr_category_warn_va_list(rb_warning_category_t category, const char *fmt, va_list args);
-#endif
 
 RBIMPL_ATTR_COLD()
 RBIMPL_ATTR_NONNULL((2))
@@ -625,18 +566,7 @@ RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 2, 3)
  * @param[in]  cat  Category e.g. deprecated.
  * @param[in]  fmt  Format specifier string compatible with rb_sprintf().
  */
-#ifdef TRUFFLERUBY
-static inline void rb_category_warn(rb_warning_category_t category, const char *fmt, ...) {
-    if (!NIL_P(ruby_verbose) && rb_warning_category_enabled_p(category)) {
-        va_list args;
-        va_start(args, fmt);
-        rb_tr_category_warn_va_list(category, fmt, args);
-        va_end(args);
-    }
-}
-#else
 void rb_category_warn(rb_warning_category_t cat, const char *fmt, ...);
-#endif
 
 RBIMPL_ATTR_NONNULL((1, 3))
 RBIMPL_ATTR_FORMAT(RBIMPL_PRINTF_FORMAT, 3, 4)
