+++ lib/cext/include/truffleruby/truffleruby-abi-version.h
@@ -25,6 +25,6 @@
 // and cannot reach the same version while having different ABI
 // ($TRUFFLERUBY_VERSION is never the same as $RUBY_VERSION).
 
-#define TRUFFLERUBY_ABI_VERSION "3.4.7.8"
+#define TRUFFLERUBY_ABI_VERSION "3.4.7.9"
 
 #endif
