+++ src/main/c/cext-trampoline/struct_by_value.c
@@ -1,0 +1,62 @@
+/* Copyright (c) 2026, TruffleRuby contributors
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived from
+ *    this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+// Returning a struct by value from Sulong to NFI is not supported.
+// So we need to declare such functions here and call a `rb_tr_` variant which takes a struct pointer instead.
+
+#include <ruby.h>
+
+void rb_tr_time_interval(VALUE num, struct timeval *result);
+
+struct timeval rb_time_interval(VALUE num) {
+  struct timeval result;
+  rb_tr_time_interval(num, &result);
+  return result;
+}
+
+void rb_tr_time_timeval(VALUE time, struct timeval *result);
+
+struct timeval rb_time_timeval(VALUE time) {
+  struct timeval result;
+  rb_tr_time_timeval(time, &result);
+  return result;
+}
+
+void rb_tr_time_timespec(VALUE time, struct timespec *result);
+
+struct timespec rb_time_timespec(VALUE time) {
+  struct timespec result;
+  rb_tr_time_timespec(time, &result);
+  return result;
+}
+
+void rb_tr_thread_wait_for(struct timeval* time);
+
+void rb_thread_wait_for(struct timeval time) {
+  rb_tr_thread_wait_for(&time);
+}
